move prometheus behind a feature flag
