+++ meilisearch-http/tests/common/server.rs
@@ -162,6 +162,7 @@ pub fn default_settings(dir: impl AsRef<Path>) -> Opt {
             max_indexing_memory: MaxMemory::unlimited(),
             ..Parser::parse_from(None as Option<&str>)
         },
+        #[cfg(feature = "metrics")]
         enable_metrics_route: true,
         ..Parser::parse_from(None as Option<&str>)
     }
